the previous commit broke something I am fixing here
Removing all the rubygems from an application broke the relation with
the security alerts, making old security alerts reapear since the status
was "Pending" as in new security alerts.

This fixes it by looking for security alerts but without the restriction
of belonging to a particular gem, and updating those security alerts in
case they exist with the gem relation.